Bullet checking function simplified And other fixes

+++ src/LineTrace.cpp
@@ -7,30 +7,30 @@
 #include <cmath>
 
 LineTrace::LineTrace(ASGE::Renderer* renderer) :
-  bullet_sprites(std::array<Sprite, 5>{ Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
-                                      Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
-                                      Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
-                                      Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
-                                      Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0)})
+  bullet_sprites(
+    std::array<Sprite, 5>{ Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
+                           Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
+                           Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
+                           Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0),
+                           Sprite(renderer, "data/images/ui/buttons/neon/blue.png", 0) })
 {
-
 }
 void LineTrace::hitCheck(float distance, ASGE::Point2D origin, float angle)
 {
   float segment_distance = distance / 250;
 
-  for(int i = 0; i < 250; i++)
+  for (size_t i = 0; i < 250; i++)
   {
-    trace_points[i].x             = origin.x + segment_distance * static_cast<float>(i) * cos(angle);
-    trace_points[i].y             = origin.y + segment_distance * static_cast<float>(i) * sin(angle);
+    trace_points[i].x = origin.x + segment_distance * static_cast<float>(i) * cos(angle);
+    trace_points[i].y = origin.y + segment_distance * static_cast<float>(i) * sin(angle);
   }
 
   hit_dist = origin.distance(hit_point);
   Logging::DEBUG(std::to_string(hit_dist));
 
-  int tracers_needed = 5;
+  size_t tracers_needed = 5;
 
-  for(int i = 0; i < tracers_needed; i++)
+  for (size_t i = 0; i < tracers_needed; i++)
   {
     bullet_sprites[i].position(trace_points[(i * 50) + 25]);
     bullet_sprites[i].dimensions(ASGE::Point2D(50, 5));
@@ -41,16 +41,16 @@ void LineTrace::hitCheck(float distance, ASGE::Point2D origin, float angle)
 }
 void LineTrace::render(ASGE::Renderer* renderer)
 {
-      for(auto& bullet_sprite : bullet_sprites)
-        {
-          bullet_sprite.render(renderer);
-        }
+  for (auto& bullet_sprite : bullet_sprites)
+  {
+    bullet_sprite.render(renderer);
+  }
 }
 
 void LineTrace::update(float dt)
 {
   tracer_timer += dt;
-  for(auto& bullet_sprite : bullet_sprites)
+  for (auto& bullet_sprite : bullet_sprites)
   {
     if (bullet_sprite.visibility())
     {
@@ -61,7 +61,7 @@ void LineTrace::update(float dt)
       }
     }
   }
-  if(has_hit)
+  if (has_hit)
   {
     trace_points.fill(0);
     has_hit = false;
